Make the comm and messagehook kernel registration functions return disposables, delete the remove* functions.
This effectively reverts dd50e935d49319b30bacd2a2c14cca935da4d8c1.

We found that a common pattern with sessions is to *not* keep a handle on the kernel (since the kernel for a session can change), but we still have the need to delete the registered handlers when the kernel changes. This is easier with a specific disposable, rather than keeping a handle to the kernel and a handle to the registered function.