Remove debugging code for editor out of sync messages.
This was put in place to try to diagnose cases when the editor value differed from the model value. In practice these differences happened when the line endings were different than the model.

These differences happened when the editor had the default line separator set (which treats \r, \n, and \r\n as line separators, but only outputs \n as a line separator). Going forward, the